backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.13
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
celery==5.4.0
redis==5.0.8
httpx==0.27.2
pytest==8.3.3
alembic==1.13.2
bcrypt==4.0.1
